+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Update Chinese (Simplified) Translations (zh_cn) (Credits: Crown-Fairy) for Village Update
 - Fix a bunch of Missing Recipes
 - Remove Polish and Tiled Red Rock from Red Rock Bricks Tags
+- Fix Maple Trapdoor and Door Render Type
 
 # 2.4.0 The Village Update -> https://github.com/Potion-Studios/Oh-The-Biomes-Weve-Gone/discussions/279
 - This is a major update not all changes are listed here, please check the discussion for more details